# LBAW1/a7_openapi.yaml
openapi: "3.0.0"
info:
  version: "1.0"
  title: "LBAW Knowledge API"
  description: "Web Resources Specification (A7) for Collaborative Q&A"
servers:
  - url: http://lbaw.fe.up.pt
    description: "production server"
...
paths:
...
  /user/{id}:
    get:
      operationId: R06
      summary: "R06: View user profile"
      description: "Show the individual user profile, will show a form is user is in its own profile. Access: USR"
      tags:
        - "M01: User Authentication"
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: "Ok. Show User Profile UI"
          /questions/store:
    post:
      operationId: R12
      summary: "R12: Post New Question Action"
      description: "Processes the new question submission. Access: PUB"
      tags:
        - "M02: Question Management"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
              required:
                - title
                - description
      responses:
        "302":
          description: "Redirect after posting a new question."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Question posted successfully. Redirect to question details."
                  value: "/allquestions"
                302Failure:
                  description: "Failed to post the question. Redirect to question form."
                  value: "/questions"
...